Rename renderer.rs to renderer_unix.rs
Replace renderer.rs with a small module that switches on OS, like the
pattern seen in status.rs.

This helps make it clear that the content formerly in renderer.rs is not
cross-platform. Types that are actually cross-platform can be relocated
to the new renderer.rs file.
